refactor: make queens the environment; hill_climber the agent
