+++ cap6635/environment/queens.py
@@ -1,5 +1,4 @@
 
-from copy import deepcopy
 from itertools import combinations
 import random
 
@@ -41,32 +40,3 @@ class NQueens:
     def successors(self):
         ''' All successor swaps '''
         return list(combinations(list(range(self._n)), 2))
-
-
-class HillClimbingQueens(NQueens):
-
-    def __init__(self, n):
-        super(HillClimbingQueens, self).__init__(n)
-        self._solution_located = False
-        self._answer = self._chess_board
-        self._cost = [self.total_tension(self._answer)]
-
-    def random_init(self):
-        super(HillClimbingQueens, self).random_init()
-        self._answer = self._chess_board
-        self._cost = [self.total_tension(self._answer)]
-
-    def climb(self):
-        for pair in super(HillClimbingQueens, self).successors():
-            new_position = deepcopy(self._answer)
-            old_pair = new_position[pair[0]]
-            new_position[pair[0]] = new_position[pair[1]]
-            new_position[pair[1]] = old_pair
-            new_cost = self.total_tension(new_position)
-            if new_cost < self._cost[-1]:
-                self._answer = new_position
-                self._cost.append(new_cost)
-                if self._cost[-1] == 0:
-                    self._solution_located = True
-                self.climb()
-        return "Failed"
