+++ nqueens.py
@@ -3,13 +3,15 @@ import os
 import random
 import sys
 
-from cap6635.agents.informed.queens import (
-    HillClimbingQueens
+from cap6635.agents.informed.hill_climbing import (
+    HillClimbing
 )
+from cap6635.environment.queens import NQueens
 from cap6635.utilities.plot import VacuumAnimator
 
 
-agent = HillClimbingQueens(32)
+board = NQueens(20)
+agent = HillClimbing(board)
 
 last_cost = agent._cost
 repetitions = 0
